Update logging configuration
Previously, exceptions were not being logged. Now, the exception type
and value will be logged and the full exception will be emailed to the
site admins.

There is also now a LOG_LEVEL environment variable that can be set to
manipulate the detail of logs ouputted by the system. The default log
level is INFO.